chore: input type
