+++ crates/core-new/src/cairo_runner/input/mod.rs
@@ -1,3 +1,2 @@
 pub mod pre_run;
 pub mod run;
-pub mod types;
