+++ crates/core-new/src/cairo_runner/input/pre_run.rs
@@ -4,16 +4,25 @@
 use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
 use serde::Serialize;
 use serde_with::serde_as;
+use starknet::core::serde::unsigned_field_element::UfeHex;
 use starknet_crypto::FieldElement;
 use std::path::PathBuf;
-
-use super::types::InputModule;
 
 #[serde_as]
 #[derive(Serialize)]
 pub struct PreRunnerInput {
     pub identified_keys_file: PathBuf,
     pub modules: Vec<InputModule>,
+}
+
+#[serde_as]
+#[derive(Serialize)]
+pub struct InputModule {
+    #[serde_as(as = "Vec<UfeHex>")]
+    pub inputs: Vec<FieldElement>,
+    /// Detail class code of the module.
+    /// This will be loaded to bootloader.
+    pub module_class: CasmContractClass,
 }
 
 impl PreRunnerInput {
