+++ crates/core-new/src/cairo_runner/input/run.rs
@@ -1,8 +1,11 @@
+use alloy_primitives::FixedBytes;
+use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
 use hdp_primitives::datalake::output::TaskFormatted;
 use hdp_provider::evm::AbstractProviderResult;
 use serde::Serialize;
-
-use super::types::InputProcessModule;
+use serde_with::serde_as;
+use starknet::core::serde::unsigned_field_element::UfeHex;
+use starknet_crypto::FieldElement;
 
 /*
     input.json file that will be passed to the processor, is generated by this struct.
@@ -27,6 +30,21 @@ pub enum InputTask {
     DatalakeCompute(TaskFormatted),
     #[serde(rename = "module")]
     Module(InputProcessModule),
+}
+
+#[serde_as]
+#[derive(Serialize)]
+pub struct InputProcessModule {
+    #[serde_as(as = "Vec<UfeHex>")]
+    pub inputs: Vec<FieldElement>,
+    /// Detail class code of the module.
+    /// This will be loaded to bootloader.
+    pub module_class: CasmContractClass,
+
+    /// The commitment of the task.
+    pub task_commitment: String,
+
+    pub task_proof: Vec<FixedBytes<32>>,
 }
 
 impl RunnerInput {
